+++ Sources/PayPalNativePayments/PayPalNativeCheckoutClient.swift
@@ -66,18 +66,20 @@ public class PayPalNativeCheckoutClient {
                     self.notifySuccess(for: result)
                 },
                 onShippingChange: { shippingChange, shippingChangeAction in
-                    shippingChangeAction.approve()
-                    
+                    let paypalShippingActions = PayPalNativePaysheetActions(shippingChangeAction)
                     switch shippingChange.type {
                     case .shippingAddress:
                         let shippingAddress = PayPalNativeShippingAddress(shippingChange.selectedShippingAddress)
-                        self.notifyShippingChange(shippingAddress: shippingAddress)
+                        self.notifyShippingChange(shippingActions: paypalShippingActions, shippingAddress: shippingAddress)
                         
                     case .shippingMethod:
-                        if let selectedShippingMethod = shippingChange.selectedShippingMethod {
-                            let shippingMethod = PayPalNativeShippingMethod(selectedShippingMethod)
-                            self.notifyShippingMethod(shippingMethod: shippingMethod)
+                        guard let selectedShippingMethod = shippingChange.selectedShippingMethod else {
+                            return
                         }
+                        self.notifyShippingMethod(
+                            shippingActions: paypalShippingActions,
+                            shippingMethod: PayPalNativeShippingMethod(selectedShippingMethod)
+                        )
                     @unknown default:
                         break // do nothing
                     }
@@ -86,7 +88,7 @@ public class PayPalNativeCheckoutClient {
                     self.notifyCancellation()
                 },
                 onError: { error in
-                    self.notifyFailure(with: error)
+                    self.notifyFailure(with: error.reason)
                 },
                 nxoConfig: nxoConfig
             )
@@ -100,10 +102,10 @@ public class PayPalNativeCheckoutClient {
         delegate?.paypal(self, didFinishWithResult: result)
     }
 
-    private func notifyFailure(with errorInfo: PayPalCheckout.ErrorInfo) {
+    private func notifyFailure(with errorDescription: String) {
         apiClient.sendAnalyticsEvent("paypal-native-payments:failed")
         
-        let error = PayPalNativePaymentsError.nativeCheckoutSDKError(errorInfo.reason)
+        let error = PayPalNativePaymentsError.nativeCheckoutSDKError(errorDescription)
         delegate?.paypal(self, didFinishWithError: error)
     }
 
@@ -112,13 +114,19 @@ public class PayPalNativeCheckoutClient {
         delegate?.paypalDidCancel(self)
     }
     
-    private func notifyShippingMethod(shippingMethod: PayPalNativeShippingMethod) {
+    private func notifyShippingMethod(
+        shippingActions: PayPalNativePaysheetActions,
+        shippingMethod: PayPalNativeShippingMethod
+    ) {
         apiClient.sendAnalyticsEvent("paypal-native-payments:shipping-method-changed")
-        shippingDelegate?.paypal(self, didShippingMethodChange: shippingMethod)
+        shippingDelegate?.paypal(self, didShippingMethodChange: shippingMethod, withAction: shippingActions)
     }
     
-    private func notifyShippingChange(shippingAddress: PayPalNativeShippingAddress) {
+    private func notifyShippingChange(
+        shippingActions: PayPalNativePaysheetActions,
+        shippingAddress: PayPalNativeShippingAddress
+    ) {
         apiClient.sendAnalyticsEvent("paypal-native-payments:shipping-address-changed")
-        shippingDelegate?.paypal(self, didShippingAddressChange: shippingAddress)
+        shippingDelegate?.paypal(self, didShippingAddressChange: shippingAddress, withAction: shippingActions)
     }
 }